+++ src/main/java/christmas/AppConfig.java
@@ -2,9 +2,9 @@ package christmas;
 
 import camp.nextstep.edu.missionutils.Console;
 import christmas.controller.ChristmasEventController;
-import christmas.model.domain.event.ChristmasEventStrategy;
+import christmas.model.domain.event.strategy.ChristmasEventStrategy;
 import christmas.model.domain.event.EventRepository;
-import christmas.model.domain.event.EventStrategy;
+import christmas.model.domain.event.strategy.EventStrategy;
 import christmas.model.domain.event.benefit.ChristmasTotalDiscount;
 import christmas.model.service.ChristmasEventService;
 import christmas.model.util.Converter;
@@ -17,20 +17,20 @@ public class AppConfig {
         return new ChristmasEventController(christmasEventService(), inputView(), outputView());
     }
 
-    private ChristmasEventService christmasEventService() {
-        return new ChristmasEventService(converter(), eventRepository(), christmasEventStrategy());
-    }
-
-    private Converter converter() {
-        return new Converter();
-    }
-
     private InputView inputView() {
         return new InputView();
     }
 
     private OutputView outputView() {
         return new OutputView();
+    }
+
+    private ChristmasEventService christmasEventService() {
+        return new ChristmasEventService(converter(), eventRepository(), christmasEventStrategy());
+    }
+
+    private Converter converter() {
+        return new Converter();
     }
 
     private EventRepository eventRepository() {
