+++ src/main/java/christmas/model/domain/event/benefit/Badge.java
@@ -1,4 +1,4 @@
-package christmas.model.domain.event;
+package christmas.model.domain.event.benefit;
 
 import java.util.Arrays;
 import java.util.Comparator;
