+++ src/main/java/christmas/model/domain/event/strategy/EventStrategy.java
@@ -1,6 +1,7 @@
-package christmas.model.domain.event;
+package christmas.model.domain.event.strategy;
 
 import christmas.model.domain.date.Date;
+import christmas.model.domain.event.EventResults;
 import christmas.model.domain.order.RequestOrders;
 
 public interface EventStrategy {
